+++ packages/attestation-service/src/attestationRoutes.ts
@@ -1,6 +1,11 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IHttpRequestContext, INoContentResponse, IRestRoute, ITag } from "@gtsc/api-models";
+import type {
+	IHttpRequestContext,
+	INoContentResponse,
+	IRestRoute,
+	ITag
+} from "@twin.org/api-models";
 import type {
 	IAttestationAttestRequest,
 	IAttestationAttestResponse,
@@ -10,10 +15,10 @@ import type {
 	IAttestationTransferResponse,
 	IAttestationVerifyRequest,
 	IAttestationVerifyResponse
-} from "@gtsc/attestation-models";
-import { ComponentFactory, Guards } from "@gtsc/core";
-import { nameof } from "@gtsc/nameof";
-import { HttpStatusCode } from "@gtsc/web";
+} from "@twin.org/attestation-models";
+import { ComponentFactory, Guards } from "@twin.org/core";
+import { nameof } from "@twin.org/nameof";
+import { HttpStatusCode } from "@twin.org/web";
 
 /**
  * The source used when communicating about these routes.