+++ packages/attestation-cli/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/attestation-cli - Changelog
+# @twin.org/attestation-cli - Changelog
 
-## v0.0.3
+## v0.0.1-next.3
 
 - Initial Release