+++ packages/attestation-connector-iota/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/attestation-connector-iota - Changelog
+# @twin.org/attestation-connector-iota - Changelog
 
-## v0.0.3
+## v0.0.1-next.3
 
 - Initial Release