+++ packages/attestation-service/docs/changelog.md
@@ -1,5 +1,5 @@
-# @gtsc/attestation-service - Changelog
+# @twin.org/attestation-service - Changelog
 
-## v0.0.3
+## v0.0.1-next.3
 
 - Initial Release